+++ src/__mocks__/pages/api/puzzles/instances/create.ts
@@ -1,6 +1,6 @@
 import * as faker from 'faker';
 import { DIFFICULTY } from '../../../../../types/global';
-import {getRandomNumberRange, getRandomString} from "../../getRandom";
+import {getRandomNumberRange, getRandomString} from "../../../../getRandom";
 
 // TODO: Ensure this list matches DB's configuration
 const DIFFICULTY_LIST = ['EASY', 'MEDIUM', 'HARD'];
