+++ src/__tests__/utils/password.test.ts
@@ -1,5 +1,5 @@
 import { checkHash, hashFunction } from '../../utils/password';
-import {getRandomNumberMin, getRandomString} from '../../__mocks__/pages/api/getRandom';
+import {getRandomNumberMin, getRandomString} from '../../__mocks__/getRandom';
 
 describe('testing the hashFunction', () => {
   it('should compare the secret hashed and without', async () => {
